==> research/02_data_validation.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

import pandas as pd

data = pd.read_csv("artifacts/data_ingestion/Sleep_Efficiency.csv")
data.head()

data.info()

Let's inspect and clean the data a bit.

data.isnull().sum()

from collections import Counter

print(Counter(data["Awakenings"]))

print(Counter(data["Caffeine consumption"]))

print(Counter(data["Alcohol consumption"]))

print(Counter(data["Exercise frequency"]))

Since we don't have a lot of data, we will just fill missing values with the most frequent value.

Also, we will modify Caffeine and Alcohol consumption columns to be binary (so 0 for no consumption and 1 for at least some).

data_filled = data.apply(lambda x: x.fillna(x.mode()[0]) if x.isnull().any() else x)

# just double check one of the columns to make sure it worked
print(Counter(data_filled["Exercise frequency"]))

data_filled.isnull().sum()

Now let's build the data class.

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    STATUS_FILE: str
    unzip_data_dir: Path
    all_schema: dict

from sleep_efficiency.constants import *
from sleep_efficiency.utils.common import read_yaml, create_directories


class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):
    
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_validation_config = DataValidationConfig(
            root_dir = config.root_dir,
            STATUS_FILE = config.STATUS_FILE,
            unzip_data_dir = config.unzip_data_dir,
            all_schema = schema
        )

        return data_validation_config

import os
from sleep_efficiency import logger

class DataValidation:
    def __init__(self, config: DataValidationConfig):
        self.config = config

    def validate_all_columns(self) -> bool:
        try:
            validation_status = None

            data = pd.read_csv(self.config.unzip_data_dir)
            all_cols = list(data.columns)

            all_schema = self.config.all_schema.keys()

            for col in all_cols:
                if col not in all_schema:
                    validation_status = False
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}")
                else:
                    validation_status = True
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}") 
            return validation_status
        except Exception as e:
            raise e



try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(config=data_validation_config)
    data_validation.validate_all_columns()
except Exception as e:
    raise e
