==> research/03_data_transformation.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

## EDA

Let's explore, clean, and figure out what transformations we need.

import pandas as pd

data = pd.read_csv("artifacts/data_ingestion/Sleep_Efficiency.csv")
data_og = data.copy() #make a copy just in case
data.head()

data.info()

We already know from the data validation notebook that we have some null values so let's deal with those.

print(data.isnull().sum())

from collections import Counter

data = data.apply(lambda x: x.fillna(x.mode()[0]) if x.isnull().any() else x)

print(data.isnull().sum())

Now let's do some take a look at other columns.

Counter(data["Gender"])

# just convert to 0 and 1
data["Gender"] = data["Gender"].map({'Female': 0, 'Male': 1})

For the bedtime and wakeup time the hour is the only important part really.

data["Bedtime"] = pd.to_datetime(data["Bedtime"]).dt.hour
data["Wakeup time"] = pd.to_datetime(data["Wakeup time"]).dt.hour

Now for sleep duration, we'll just leave it as an int.

Counter(data["Sleep duration"])

The Sleep efficiency, REM sleep percentage, Deep sleep percentage, and Light sleep percentage are all columns really target variables. We only care about sleep efficiency here but could add the others later.

data.columns

Counter(data["Awakenings"])

Let's leave awakenings as is.

Counter(data["Caffeine consumption"])

Counter(data["Alcohol consumption"])

These are in mg. These values don't seem realistic, let's just change it to true or false on whether or not caffiene was consumed the day before. Let's do the same thing for alcohol.

data["Caffeine consumption"] = data["Caffeine consumption"].apply(lambda x: 1 if x > 0 else 0)
Counter(data["Caffeine consumption"])

data["Alcohol consumption"] = data["Alcohol consumption"].apply(lambda x: 1 if x > 0 else 0)
Counter(data["Alcohol consumption"])

Last two columns, smoking status just needs to be converted into binary int column.

data["Smoking status"][0:5]

data["Smoking status"] = data["Smoking status"].map({'Yes': 1, 'No': 0})
data["Smoking status"][0:5]

Now looking at the exercise frequency below, we can see exercising 0 or 1 time per week are common and both would be considered infrequent, so let's lump them together. Then combine the rest and make this a binary column.

Counter(data["Exercise frequency"])

data["Exercise frequency"] = data["Exercise frequency"].apply(lambda x: 1 if x > 1 else 0)
Counter(data["Exercise frequency"])

data.head()

We can drop the target columns that we don't need for now, and also the ID column.

data = data.drop(columns=["ID", "REM sleep percentage", "Deep sleep percentage", "Light sleep percentage"], axis=1)
data.head()

Now back to modularizing the code.

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    data_path: Path

from sleep_efficiency.constants import *
from sleep_efficiency.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):
    
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        data_transformation_config = DataTransformationConfig(
            root_dir=config.root_dir,
            data_path=config.data_path,
        )

        return data_transformation_config

import os
from sleep_efficiency import logger
from sklearn.model_selection import train_test_split
import pandas as pd

class DataTransformation:
    def __init__(self, config: DataTransformationConfig):
        self.config = config

    def train_test_splitting(self):
        data = pd.read_csv(self.config.data_path)
        
        # cleaning and preprocessing data
        data = data.apply(lambda x: x.fillna(x.mode()[0]) if x.isnull().any() else x) # filling all null values with most frequent

        # converting some columns to binary 0 and 1
        data["Gender"] = data["Gender"].map({'Female': 0, 'Male': 1})
        data["Caffeine consumption"] = data["Caffeine consumption"].apply(lambda x: 1 if x > 0 else 0)
        data["Alcohol consumption"] = data["Alcohol consumption"].apply(lambda x: 1 if x > 0 else 0)
        data["Smoking status"] = data["Smoking status"].map({'Yes': 1, 'No': 0})
        data["Exercise frequency"] = data["Exercise frequency"].apply(lambda x: 1 if x > 1 else 0)

        #converting time columns to just the hour
        data["Bedtime"] = pd.to_datetime(data["Bedtime"]).dt.hour
        data["Wakeup time"] = pd.to_datetime(data["Wakeup time"]).dt.hour

        # dropping other potential target columns
        data = data.drop(columns=["ID", "REM sleep percentage", "Deep sleep percentage", "Light sleep percentage"], axis=1)
        
        train, test = train_test_split(data)
        
        train.to_csv(os.path.join(self.config.root_dir, "train.csv"), index=False)
        test.to_csv(os.path.join(self.config.root_dir, "test.csv"), index=False)

        logger.info("Splitting data into training and testing sets")
        logger.info(train.shape)
        logger.info(test.shape)

        print(train.shape)
        print(test.shape)

try:
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformation(config = data_transformation_config)
    data_transformation.train_test_splitting()
except Exception as e:
    raise e
