==> research/04_model_trainer.ipynb <==
# Model Trainer Notebook

Modifying file path.

import os

%pwd

os.chdir("../")

%pwd

## Model Tuning

import pandas as pd
from sklearn.linear_model import ElasticNet

train_data = pd.read_csv("artifacts/data_transformation/train.csv")
test_data = pd.read_csv("artifacts/data_transformation/test.csv")

train_x = train_data.drop(['Sleep efficiency'], axis=1)
test_x = test_data.drop(['Sleep efficiency'], axis=1)
train_y = train_data[['Sleep efficiency']]
test_y = test_data[['Sleep efficiency']]

lr = ElasticNet(alpha=0.4, l1_ratio=0.8, random_state=42)
lr.fit(train_x, train_y)

data = test_x.iloc[5]
data

feature_names = list(test_x.columns)
feature_names

data_df = pd.DataFrame([data], columns=feature_names)

lr.predict(data_df)

for i in range(0, 15):
    data = test_x.iloc[i]
    data_df = pd.DataFrame([data], columns=feature_names)
    print(lr.predict(data_df))

ElasticNet is terrible, let's try XGBoost.

import xgboost as xgb

model = xgb.XGBRegressor(objective='reg:squarederror', random_state=42)

model.fit(train_x, train_y)

# Predict using the test data
predictions = model.predict(test_x)

print("First 10 predictions:", predictions[:10])

test_y[:10]

This already looks so much better. Let's do grid search and run with the best parameters from that.

from sklearn.model_selection import GridSearchCV

model = xgb.XGBRegressor(objective='reg:squarederror', random_state=42)

# Define the parameter grid
param_grid = {
    'n_estimators': [None, 50, 100],
    'learning_rate': [0.05, 0.1, 0.2],
    'max_depth': [None, 3, 5],
    'subsample': [0.7, 0.8, 0.9],
    'colsample_bytree': [0.2, 0.4, 0.6, 0.8]
}

# Set up the GridSearchCV
grid_search = GridSearchCV(estimator=model, param_grid=param_grid, cv=5, scoring='neg_mean_squared_error', verbose=1, n_jobs=-1)

# Fit the model to the training data
grid_search.fit(train_x, train_y)

from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

best_model = grid_search.best_estimator_
print("Best parameters found: ", grid_search.best_params_)

predictions = best_model.predict(test_x)
print("First 10 predictions:", predictions[:10])
print("First 10 true values:", test_y[:10])

mse = mean_squared_error(test_y, predictions)
rmse = np.sqrt(mse)
r2 = r2_score(test_y, predictions)

print("Mean Squared Error: ", mse)
print("Root Mean Squared Error: ", rmse)
print("R² Score: ", r2)

R2 score is not great but not bad. We will run with it for now.

## Modularizing Model Training

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class ModelTrainerConfig:
    root_dir: Path
    train_data_path: Path
    test_data_path: Path
    model_name: str
    n_estimators: float
    learning_rate: float
    max_depth: float
    subsample: float
    colsample_bytree: float
    target_column: str

from sleep_efficiency.constants import *
from sleep_efficiency.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):
    
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_model_trainer_config(self) -> ModelTrainerConfig:
        config = self.config.model_trainer
        params = self.params.XGBoost
        schema = self.schema.TARGET_COLUMN

        create_directories([config.root_dir])

        model_trainer_config = ModelTrainerConfig(
            root_dir = config.root_dir,
            train_data_path = config.train_data_path,
            test_data_path = config.test_data_path,
            model_name = config.model_name,
            n_estimators = params.n_estimators,
            learning_rate = params.learning_rate,
            max_depth = params.max_depth,
            subsample = params.subsample,
            colsample_bytree = params.colsample_bytree,
            target_column = schema.name
        )
        
        return model_trainer_config

import pandas as pd
import os
from sleep_efficiency import logger
import xgboost as xgb
import joblib

class ModelTrainer:
    def __init__(self, config: ModelTrainerConfig):
        self.config = config

    def train(self):
        train_data = pd.read_csv(self.config.train_data_path)
        test_data = pd.read_csv(self.config.test_data_path)

        train_x = train_data.drop([self.config.target_column], axis=1)
        test_x = test_data.drop([self.config.target_column], axis=1)
        train_y = train_data[[self.config.target_column]]
        test_y = test_data[[self.config.target_column]]

        model = xgb.XGBRegressor(
            objective='reg:squarederror',
            random_state=42,
            n_estimators=self.config.n_estimators,
            learning_rate=self.config.learning_rate,
            max_depth=self.config.max_depth,
            subsample=self.config.subsample,
            colsample_bytree=self.config.colsample_bytree
        )
        model.fit(train_x, train_y)

        joblib.dump(model, os.path.join(self.config.root_dir, self.config.model_name))


try:
    config = ConfigurationManager()
    model_trainer_config = config.get_model_trainer_config()
    model_trainer_config = ModelTrainer(config=model_trainer_config)
    model_trainer_config.train()
except Exception as e:
    raise e
